+++ CHANGELOG.md
@@ -16,12 +16,13 @@
 + [x] Table: add column rownum. This column shows when attribute `data-rownum="true"`
 + [x] Table: add column row check. This column shows when attribute `data-check="true"`
 + [x] Table: fix create internal heads when header defining in html and data loaded from json 
-+ [x] Table: fix create internal foots when header defining in html and data loaded from json
 + [x] Table: add `data-locale` attribute
 + [x] Table: add table inspector to configure columns view
 + [x] Table: add methods `openInspector(true|false)`, `toggleInspector()` to show/hide table inspector
 + [x] Table: add attribute `data-inpector-save-mode="client|server"`
 + [x] Table: add attribute `data-inpector-save-path="storage key|url"`
++ [x] Table: add methods `getFilteredItems()`, `getSelectedItems()`, `getStoredKeys()`
++ [x] Table: add events `onDrawCell`, `onAppendCell`, `onAppendRow`, `onViewSave`, `onViewGet`, `onCheckDraw`
 
 ### 4.2.14
 + [x] Table: fix pagination calculator when rows count changed